Refactor : 인터페이스 삭제
확장 가능성이 없는 클래스 인터페이스 삭제
